Funding 수정

+++ src/main/java/re21/ieun/funding/dto/FundingUpcyclingDto.java
@@ -15,6 +15,6 @@ public class FundingUpcyclingDto {
     private long upcyclingId;
 
     @Positive
-    private long quantity;
+    private int quantity;
 
 }
